refactor: exam page
