+++ src/public/pages/authorized-user/exam/exam.js
@@ -9,81 +9,79 @@ const url = `http://localhost:4000`;
 
 function ExamPage() {
   const element = document.createElement("div");
+  const header = Header();
+  const headerWithoutMenu = HeaderWithoutMenu();
+
+  element.appendChild(header);
 
   element.style.height = "100vh";
   element.style.display = "flex";
   element.style.flexDirection = "column";
   element.style.alignItems = "center";
-  const nowDate = new Date();
 
-  function renderInitialPageNewExam(simulados) {
-    fetch(url + "/api/points", {
-      method: "GET",
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        const container = document.createElement("div");
-        container.id = "containerInitial";
-        container.innerHTML = `
+  function InitialPageStartNewExam() {
+    const container = document.createElement("div");
+    container.id = "containerInitial";
+    container.innerHTML = `
             <main>
                 <h1>SIMULADO ENEM</h1>
-                <p>Um mini simulado do ENEM com 12 questões e duração aproximada de 30 minutos. O nível de dificuldade é ${data.level}</p>
+                <p>Um mini simulado do ENEM com 12 questões e duração aproximada de 30 minutos. O nível de dificuldade é <span id="examLevel">2</span></p>
                 <button id="startButton" class="button">INICIAR</button>
-                <span id="history">Ver historico de simulados</span>
             </main>
         `;
 
-        let viewportWidth = window.innerWidth;
-        const examImage = document.createElement("img");
-        let addedImage = false;
-        examImage.src = "../../images/site/Simulado.png";
-        examImage.id = "examImage";
-        examImage.alt = "Estudante Ilustração";
+    let viewportWidth = window.innerWidth;
+    const examImage = document.createElement("img");
+    let addedImage = false;
+    examImage.src = "../../images/site/Simulado.png";
+    examImage.id = "examImage";
+    examImage.alt = "Estudante Ilustração";
 
-        if (viewportWidth > 1100) {
-          addedImage = true;
-          container.appendChild(examImage);
-        }
+    if (viewportWidth > 1100) {
+      addedImage = true;
+      container.appendChild(examImage);
+    }
 
-        window.addEventListener("resize", () => {
-          viewportWidth = window.innerWidth;
-          if (viewportWidth > 1100 && !addedImage) {
-            addedImage = true;
-            container.appendChild(examImage);
-          } else if (viewportWidth < 1100 && addedImage) {
-            container.removeChild(examImage);
-            addedImage = false;
-          }
+    window.addEventListener("resize", () => {
+      viewportWidth = window.innerWidth;
+      if (viewportWidth > 1100 && !addedImage) {
+        addedImage = true;
+        container.appendChild(examImage);
+      } else if (viewportWidth < 1100 && addedImage) {
+        container.removeChild(examImage);
+        addedImage = false;
+      }
+    });
+
+    container.style.padding = "0px 20px";
+
+    const startButton = container.querySelector("#startButton");
+
+    element.appendChild(container);
+
+    startButton.addEventListener("click", () => {
+      fetch(`${url}/api/exam`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      })
+        .then((response) => response.json())
+        .then((data) => {
+          renderQuestionsPage(data);
         });
-
-        container.style.padding = "0px 20px";
-
-        const startButton = container.querySelector("#startButton");
-        const header = Header();
-
-        element.appendChild(header);
-        element.appendChild(container);
-
-        startButton.addEventListener("click", () => {
-          fetch(`${url}/api/exam`, {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-          })
-            .then((response) => response.json())
-            .then((data) => {
-              renderQuestionsPage(data);
-            });
-        });
-        const history = container.querySelector("#history");
-        history.addEventListener("click", () => {
-          renderHistoryPage(simulados);
-        });
-      });
+    });
+    async function fetchUserLevel() {
+      const response = await fetch(`${url}/api/points`);
+      const data = await response.json();
+      const level = container.querySelector("#examLevel");
+      level.textContent = data.level;
+    }
+    fetchUserLevel();
+    return container;
   }
 
-  function renderInitialPageContinue(simulado) {
+  function InitialPageContinueExam(exam) {
     const container = document.createElement("div");
     container.id = "containerInitial";
     container.innerHTML = `
@@ -118,14 +116,12 @@ function ExamPage() {
     });
 
     const startButton = container.querySelector("#startButton");
-    const header = Header();
-
-    element.appendChild(header);
-    element.appendChild(container);
 
     startButton.addEventListener("click", () => {
-      renderQuestionsPage(simulado);
+      renderQuestionsPage(exam);
     });
+
+    return container;
   }
 
   function renderQuestionsPage(simulado) {
@@ -133,9 +129,7 @@ function ExamPage() {
 
     element.innerHTML = ``;
 
-    const header = HeaderWithoutMenu();
-
-    element.appendChild(header);
+    element.appendChild(headerWithoutMenu);
 
     const containerExam = document.createElement("div");
     containerExam.id = "containerExam";
@@ -222,6 +216,7 @@ function ExamPage() {
         .then((data) => {
           renderResultPage(data);
         });
+      window.scrollTo({ behavior: "smooth", top: 0 });
     }
   }
 
@@ -231,7 +226,6 @@ function ExamPage() {
     element.innerHTML = "";
     const header = Header();
     header.style.width = "100vw";
-
     element.appendChild(header);
 
     const containerResults = document.createElement("div");
@@ -243,187 +237,29 @@ function ExamPage() {
     });
   }
 
-  function renderHistoryPage(simulados) {
-    element.innerHTML = "";
-    const header = Header();
-    header.style.width = "100vw";
-
-    element.appendChild(header);
-
-    const containerHistory = document.createElement("div");
-    containerHistory.id = "containerHistory";
-    element.appendChild(containerHistory);
-
-    containerHistory.style.width = "95%";
-    containerHistory.style.maxWidth = "1200px";
-    containerHistory.style.display = "flex";
-    containerHistory.style.flexDirection = "column";
-    containerHistory.style.gap = "10px";
-    containerHistory.style.alignItems = "center";
-    containerHistory.style.marginTop = "10px";
-
-    simulados.forEach((simulado) => {
-      console.log(simulado);
-      if (!simulado.done) {
-        return;
-      }
-      containerHistory.appendChild(examElement(simulado));
+  async function fetchExamsHistory() {
+    const response = await fetch(`${url}/api/exam`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
     });
+    const data = await response.json();
+    let notDoneExam;
+    data.forEach((exam) => {
+      if (!exam.done) {
+        notDoneExam = exam;
+      }
+    });
+
+    if (!notDoneExam) {
+      element.appendChild(InitialPageStartNewExam());
+    } else {
+      element.appendChild(InitialPageContinueExam(notDoneExam));
+    }
   }
 
-  function examElement(exam) {
-    const examElement = document.createElement("div");
-    const examInformations = document.createElement("div");
-    const examImportantInformation = document.createElement("div");
-    const examExtraInformation = document.createElement("div");
-    const examResults = document.createElement("div");
-
-    // Styling exam container
-    examElement.style.width = "100%";
-    examElement.style.border = "1px solid #ddd";
-    examElement.style.borderRadius = "8px";
-    examElement.style.padding = "10px";
-    examElement.style.margin = "10px 0";
-    examElement.style.backgroundColor = "#ffff";
-    examElement.style.boxShadow = "0 2px 4px rgba(0, 0, 0, 0.1)";
-
-    // Styling exam information section
-    examImportantInformation.style.textAlign = "center";
-    examImportantInformation.style.cursor = "pointer";
-    examImportantInformation.style.padding = "10px";
-    examImportantInformation.style.backgroundColor = "#0B2072";
-    examImportantInformation.style.color = "white";
-    examImportantInformation.style.borderRadius = "5px";
-    examImportantInformation.style.fontWeight = "bold";
-    examImportantInformation.style.display = "flex";
-    examImportantInformation.style.justifyContent = "space-around";
-
-    examExtraInformation.style.display = "flex";
-    examExtraInformation.style.justifyContent = "space-around";
-    examExtraInformation.style.gap = "10px";
-    examExtraInformation.style.marginTop = "10px";
-    examExtraInformation.style.flexWrap = "wrap";
-
-    const dateLimit = new Date(exam.limit_time);
-    const dateDone = new Date(exam.done_time_at);
-    const date = dateDone > dateLimit ? dateLimit : dateDone;
-
-    let totalQuestions = exam.questions.length;
-    let correct = 0;
-
-    exam.questions.forEach((question) => {
-      question.alternatives.forEach((alternative) => {
-        if (alternative.is_correct && alternative.id === question.answer_id) {
-          correct += 1;
-        }
-      });
-    });
-
-    examImportantInformation.innerHTML = `
-      <span>Feito Em: ${String(date.getDate()).padStart(2, "0")}/${String(
-      date.getMonth() + 1
-    ).padStart(2, "0")}/${date.getFullYear()} - ${String(
-      date.getHours()
-    ).padStart(2, "0")}:${String(date.getMinutes()).padStart(2, "0")}</span>
-      <span>Acertos totais: ${correct}/${totalQuestions}</span>
-    `;
-
-    examInformations.appendChild(examImportantInformation);
-    examInformations.appendChild(examExtraInformation);
-    examElement.appendChild(examInformations);
-    examElement.appendChild(examResults);
-
-    // Styling results container
-    examResults.style.display = "none";
-    examResults.style.padding = "10px";
-    examResults.style.marginTop = "10px";
-    examResults.style.borderTop = "1px solid #ddd";
-    examResults.style.backgroundColor = "#ffffff";
-    examResults.style.borderRadius = "5px";
-
-    let disciplines = {};
-
-    exam.questions.forEach((question) => {
-      if (disciplines[question.discipline]) {
-        disciplines[question.discipline].total += 1;
-        question.alternatives.forEach((alternative) => {
-          if (alternative.is_correct && alternative.id === question.answer_id) {
-            disciplines[question.discipline].correct += 1;
-          }
-        });
-      } else {
-        disciplines[question.discipline] = {
-          total: 1,
-          correct: 0,
-        };
-        question.alternatives.forEach((alternative) => {
-          if (alternative.is_correct && alternative.id === question.answer_id) {
-            disciplines[question.discipline].correct += 1;
-          }
-        });
-      }
-    });
-
-    Object.keys(disciplines).forEach((discipline) => {
-      const disciplineElement = document.createElement("div");
-      disciplineElement.style.display = "flex";
-      disciplineElement.style.flexDirection = "column";
-      disciplineElement.style.gap = "10px";
-      disciplineElement.style.alignItems = "center";
-      disciplineElement.style.padding = "10px";
-      disciplineElement.style.backgroundColor = "#f4f4f4";
-      disciplineElement.style.borderRadius = "7px";
-      disciplineElement.style.boxShadow = "2px 2px 10px rgba(0, 0, 0, 0.1)";
-      disciplineElement.innerHTML = `<div>
-        <span>${discipline
-          .split("-")
-          .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-          .join(" ")}</span>
-        <span>${disciplines[discipline].correct}/${
-        disciplines[discipline].total
-      }</span>
-      </div>`;
-      examExtraInformation.appendChild(disciplineElement);
-    });
-    examInformations.addEventListener("click", () => {
-      if (examResults.style.display === "none") {
-        examResults.style.display = "flex";
-        examResults.style.flexDirection = "column";
-        examResults.style.gap = "10px";
-        examResults.innerHTML = "";
-        exam.questions.forEach((question) => {
-          examResults.appendChild(QuestionElementResultHistory(question));
-        });
-      } else {
-        examResults.style.display = "none";
-      }
-    });
-    return examElement;
-  }
-
-  fetch(`${url}/api/exam`, {
-    method: "GET",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  })
-    .then((response) => response.json())
-    .then((data) => {
-      let simulado;
-      data.forEach((exam) => {
-        if (!exam.done) {
-          simulado = exam;
-        }
-      });
-      if (!simulado) {
-        data.sort(function (a, b) {
-          return new Date(b.limit_time) - new Date(a.limit_time);
-        });
-        renderInitialPageNewExam(data);
-      } else {
-        renderInitialPageContinue(simulado);
-      }
-    });
+  fetchExamsHistory();
 
   return element;
 }
